refactor/moved folder
renamed /routes to /Quiz and moved it into /container.
Removed an unnecessary App.css import
