+++ src/container/Quiz/Quiz.jsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import "../App.css";
 import "./Quiz.scss";
 
 function Quiz() {
