+++ src/container/Quiz/Quiz.scss
