--- src/container/QuizPage/QuizPage.scss
@@ -1,13 +1,0 @@
-.quizPage__section {
-
-    .quizPage__house-sigil {
-        width: 30%;
-    }
-
-    .quizPage__btn-div {
-        display: flex;
-        flex-direction: column;
-
-        
-    }
-}
